Move resource unit hook context to apiserver
This is part of extracting the resources API facdes to the apiserver
package (like all other facades).
