+++ apiserver/resourceshookcontext/unitfacade_test.go
@@ -1,7 +1,7 @@
-// Copyright 2016 Canonical Ltd.
+// Copyright 2017 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package server_test
+package resourceshookcontext_test
 
 import (
 	"github.com/juju/errors"
@@ -11,9 +11,9 @@ import (
 
 	"github.com/juju/juju/apiserver/common"
 	"github.com/juju/juju/apiserver/params"
+	"github.com/juju/juju/apiserver/resourceshookcontext"
 	"github.com/juju/juju/resource"
 	"github.com/juju/juju/resource/api"
-	"github.com/juju/juju/resource/api/private/server"
 	"github.com/juju/juju/resource/resourcetesting"
 )
 
@@ -34,7 +34,7 @@ func (s *UnitFacadeSuite) SetUpTest(c *gc.C) {
 func (s *UnitFacadeSuite) TestNewUnitFacade(c *gc.C) {
 	expected := &stubUnitDataStore{Stub: s.stub}
 
-	uf := server.NewUnitFacade(expected)
+	uf := resourceshookcontext.NewUnitFacade(expected)
 
 	s.stub.CheckNoCalls(c)
 	c.Check(uf.DataStore, gc.Equals, expected)
@@ -49,7 +49,7 @@ func (s *UnitFacadeSuite) TestGetResourceInfoOkay(c *gc.C) {
 	store.ReturnListResources = resource.ServiceResources{
 		Resources: []resource.Resource{res1, res2},
 	}
-	uf := server.UnitFacade{DataStore: store}
+	uf := resourceshookcontext.UnitFacade{DataStore: store}
 
 	results, err := uf.GetResourceInfo(params.ListUnitResourcesArgs{
 		ResourceNames: []string{"spam", "eggs"},
@@ -72,7 +72,7 @@ func (s *UnitFacadeSuite) TestGetResourceInfoEmpty(c *gc.C) {
 	store.ReturnListResources = resource.ServiceResources{
 		Resources: []resource.Resource{opened.Resource},
 	}
-	uf := server.UnitFacade{DataStore: store}
+	uf := resourceshookcontext.UnitFacade{DataStore: store}
 
 	results, err := uf.GetResourceInfo(params.ListUnitResourcesArgs{
 		ResourceNames: []string{},
@@ -91,7 +91,7 @@ func (s *UnitFacadeSuite) TestGetResourceInfoNotFound(c *gc.C) {
 	store.ReturnListResources = resource.ServiceResources{
 		Resources: []resource.Resource{opened.Resource},
 	}
-	uf := server.UnitFacade{DataStore: store}
+	uf := resourceshookcontext.UnitFacade{DataStore: store}
 
 	results, err := uf.GetResourceInfo(params.ListUnitResourcesArgs{
 		ResourceNames: []string{"eggs"},
