+++ component/all/resource.go
@@ -5,7 +5,6 @@ package all
 
 import (
 	"os"
-	"reflect"
 
 	jujucmd "github.com/juju/cmd"
 	"github.com/juju/errors"
@@ -23,7 +22,6 @@ import (
 	"github.com/juju/juju/resource/api/client"
 	privateapi "github.com/juju/juju/resource/api/private"
 	internalclient "github.com/juju/juju/resource/api/private/client"
-	internalserver "github.com/juju/juju/resource/api/private/server"
 	"github.com/juju/juju/resource/cmd"
 	"github.com/juju/juju/resource/context"
 	contextcmd "github.com/juju/juju/resource/context/cmd"
@@ -144,8 +142,6 @@ func (r resources) registerPublicCommands() {
 	})
 }
 
-// TODO(katco): This seems to be common across components. Pop up a
-// level and genericize?
 func (r resources) registerHookContext() {
 	if markRegistered(resource.ComponentName, "hook-context") == false {
 		return
@@ -165,7 +161,6 @@ func (r resources) registerHookContext() {
 	)
 
 	r.registerHookContextCommands()
-	r.registerHookContextFacade()
 	r.registerUnitDownloadEndpoint()
 }
 
@@ -190,16 +185,7 @@ func (r resources) registerHookContextCommands() {
 	)
 }
 
-func (r resources) registerHookContextFacade() {
-	common.RegisterHookContextFacade(
-		context.HookContextFacade,
-		internalserver.FacadeVersion,
-		r.newHookContextFacade,
-		reflect.TypeOf(&internalserver.UnitFacade{}),
-	)
-
-}
-
+// XXX
 func (r resources) registerUnitDownloadEndpoint() {
 	common.RegisterAPIModelEndpoint(privateapi.HTTPEndpointPattern, apihttp.HandlerSpec{
 		Constraints: apihttp.HandlerConstraints{
@@ -211,29 +197,9 @@ func (r resources) registerUnitDownloadEndpoint() {
 	})
 }
 
-// resourcesUnitDatastore is a shim to elide serviceName from
-// ListResources.
-type resourcesUnitDataStore struct {
-	resources corestate.Resources
-	unit      *corestate.Unit
-}
-
-// ListResources implements resource/api/private/server.UnitDataStore.
-func (ds *resourcesUnitDataStore) ListResources() (resource.ServiceResources, error) {
-	return ds.resources.ListResources(ds.unit.ApplicationName())
-}
-
-func (r resources) newHookContextFacade(st *corestate.State, unit *corestate.Unit) (interface{}, error) {
-	res, err := st.Resources()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	return internalserver.NewUnitFacade(&resourcesUnitDataStore{res, unit}), nil
-}
-
 func (r resources) newUnitFacadeClient(unitName string, caller base.APICaller) (context.APIClient, error) {
 
-	facadeCaller := base.NewFacadeCallerForVersion(caller, context.HookContextFacade, internalserver.FacadeVersion)
+	facadeCaller := base.NewFacadeCallerForVersion(caller, context.HookContextFacade, 1)
 	httpClient, err := caller.HTTPClient()
 	if err != nil {
 		return nil, errors.Trace(err)
